Versión 1.4 ¡Cuántos likes!

+++ ProyectoIntegrador-Final/src/main/java/com/integrador/repository/ComentarioRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ComentarioRepository extends JpaRepository<Comentario, Long> {
+public interface ComentarioRepository extends JpaRepository<Comentario, Integer> {
     List<Comentario> findByLibro(Libro libro);
 }
